やねうら王V5.40 2020/12/22 キャッチアップ

+++ source/extra/test_cmd.cpp
@@ -3464,7 +3464,7 @@ void gen_mate(Position& pos, istringstream& is)
 	};
 
 	auto threads = make_unique<std::thread[]>(thread_num);
-	for (int i = 0; i < thread_num; ++i)
+	for (size_t i = 0; i < thread_num; ++i)
 		threads[i] = std::thread(worker,i);
 
 	while (generated_count < loop_max)
@@ -3473,7 +3473,7 @@ void gen_mate(Position& pos, istringstream& is)
 		sync_cout << generated_count << sync_endl;
 	}
 
-	for (int i = 0; i < thread_num; ++i)
+	for (size_t i = 0; i < thread_num; ++i)
 		threads[i].join();
 
 	sync_cout << "..done" << sync_endl;
